function-source/requirements.txt
# Para interagir com os buckets do Google Cloud Storage
google-cloud-storage

# Para processamento de imagens
Pillow

# Para converter DOCX para PDF
docx2pdf

# Para processamento de vídeo (extrair thumbnail)
ffmpeg-python